Add files via upload
The following code is of an offline dictionary and synonym finder. The code is written in JAVA and can only be run on Netbeans idle because it uses netbeans forms for GUI. 
The data structures used for this code are BK-Tree (to store the dictionary), Graphs (to store the synonyms) and Queue (for printing the suggestions of an incorrect word).